# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /runs:
    post:
      operationId: submitRun
      tags: [runs]
      summary: Submit a completed game run
      description: >
        Records one run. Idempotent on idempotencyKey so a kiosk reload or a
        flaky network never double-counts. Points are clamped server-side to the
        per-game cap. A player's score for a game is their best single run that
        day, so a weaker replay never lowers a standing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunInput"
      responses:
        "200":
          description: Run recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResult"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorBody"
  /runs/progress:
    post:
      operationId: saveRunProgress
      tags: [runs]
      summary: Persist in-flight run progress
      description: >
        Stores the current level, banked points and lifeline state for a run in
        progress so a kiosk browser reload resumes rather than losing the run.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunProgressInput"
      responses:
        "200":
          description: Progress saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunProgress"
  /runs/progress/{idempotencyKey}:
    get:
      operationId: getRunProgress
      tags: [runs]
      summary: Resume in-flight run progress
      parameters:
        - name: idempotencyKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Progress found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunProgress"
        "404":
          description: No progress stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorBody"
...
  /spoof/detect:
    post:
      operationId: detectSpoof
      tags: [spoof]
      summary: Run the liveness detector against an uploaded image
      description: >
        Sends one image to the detector at the requested strictness level and
        returns the verdict, the signals that fired and the heatmap regions to
        overlay. Uploads are retained for at most 24 hours.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpoofAttemptInput"
      responses:
        "200":
          description: Detector verdict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectorVerdict"
        "400":
          description: Rejected upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorBody"
...
components:
...
  schemas:
...
    ApiErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
        field:
          type: ["string", "null"]
    GameKey:
      type: string
      enum:
        - spot_the_fraud
        - spoof_the_system
        - fraud_detective
    RunSource:
      type: string
      enum: [kiosk, phone]
    LeaderboardScope:
      type: string
      enum: [today, cumulative]
...
    GameScore:
      type: object
      required: [game, points, cap, played]
      properties:
        game:
          $ref: "#/components/schemas/GameKey"
        points:
          type: number
          description: Best single run for this game within the scope.
        cap:
          type: number
        played:
          type: boolean
    PlayerStanding:
      type: object
      required: [playerId, scope, scores, bonus, total, maxTotal, rank, playedAllThree]
      properties:
        playerId:
          type: string
        scope:
          $ref: "#/components/schemas/LeaderboardScope"
        scores:
          type: array
          items:
            $ref: "#/components/schemas/GameScore"
        bonus:
          type: number
          description: The 40-point Fraud Fighter bonus, derived - never stored.
        total:
          type: number
        maxTotal:
          type: number
        rank:
          type: ["number", "null"]
        playedAllThree:
          type: boolean
    RunInput:
      type: object
      required: [playerId, game, points, source, idempotencyKey]
      properties:
        playerId:
          type: string
        game:
          $ref: "#/components/schemas/GameKey"
        points:
          type: number
          minimum: 0
        source:
          $ref: "#/components/schemas/RunSource"
        idempotencyKey:
          type: string
          minLength: 8
        detail:
          type: object
          additionalProperties: true
          description: >
            Per-game breakdown. Spot the Fraud records each level outcome and lifeline
            use; Spoof the System records the three attempts; Fraud Detective
            records per-case guesses and the bonus round.
    RunResult:
      type: object
      required: [runId, game, pointsSubmitted, pointsRecorded, isPersonalBest, duplicate, standing]
      properties:
        runId:
          type: string
        game:
          $ref: "#/components/schemas/GameKey"
        pointsSubmitted:
          type: number
        pointsRecorded:
          type: number
          description: Submitted points after clamping to the per-game cap.
        isPersonalBest:
          type: boolean
          description: True when this run beat the player's previous best for the day.
        duplicate:
          type: boolean
          description: True when this idempotency key had already been recorded.
        standing:
          $ref: "#/components/schemas/PlayerStanding"
    RunProgressInput:
      type: object
      required: [idempotencyKey, playerId, game, state]
      properties:
        idempotencyKey:
          type: string
        playerId:
          type: string
        game:
          $ref: "#/components/schemas/GameKey"
        state:
          type: object
          additionalProperties: true
    RunProgress:
      type: object
      required: [idempotencyKey, playerId, game, state, updatedAt]
      properties:
        idempotencyKey:
          type: string
        playerId:
          type: string
        game:
          $ref: "#/components/schemas/GameKey"
        state:
          type: object
          additionalProperties: true
        updatedAt:
          type: string
          format: date-time
...
    SpoofAttemptInput:
      type: object
      required: [playerId, level, image, mimeType]
      properties:
        playerId:
          type: string
        level:
          type: number
          minimum: 1
          maximum: 3
        image:
          type: string
          description: Base64-encoded image bytes, without a data URL prefix.
        mimeType:
          type: string
          description: One of image/jpeg, image/png or image/webp.
        fileName:
          type: string
    HeatmapRegion:
      type: object
      required: [x, y, w, h, intensity]
      description: Normalised 0-1 box over a suspect area of the image.
      properties:
        x:
          type: number
        y:
          type: number
        w:
          type: number
        h:
          type: number
        intensity:
          type: number
    DetectorSignal:
      type: object
      required: [name, verdict, score]
      properties:
        name:
          type: string
          description: e.g. "Frequency-domain artefacts"
        verdict:
          type: string
          enum: [synthetic, authentic, inconclusive]
        score:
          type: number
          description: 0-1
    DetectorVerdict:
      type: object
      required: [fooled, confidence, signals, heatmapRegions, latencyMs, level, uploadId]
      properties:
        fooled:
          type: boolean
        confidence:
          type: number
          description: Detector confidence that the image is a spoof, 0-1.
        signals:
          type: array
          items:
            $ref: "#/components/schemas/DetectorSignal"
        heatmapRegions:
          type: array
          items:
            $ref: "#/components/schemas/HeatmapRegion"
        latencyMs:
          type: number
        level:
          type: number
        uploadId:
          type: string
        detectorName:
          type: string
...