# lib/api-zod/src/generated/types/runProgress.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameKey } from './gameKey';
import type { RunProgressState } from './runProgressState';

export interface RunProgress {
  idempotencyKey: string;
  playerId: string;
  game: GameKey;
  state: RunProgressState;
  updatedAt: Date;
}