# lib/api-zod/src/generated/types/runSource.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type RunSource = typeof RunSource[keyof typeof RunSource];


export const RunSource = {
  kiosk: 'kiosk',
  phone: 'phone',
} as const;